refactor: replace ConvertKey command with DerivePublicKey
- Removed the ConvertKey command and its associated implementation from the CLI.
- Introduced the DerivePublicKey command to handle the conversion of private keys to public keys, maintaining similar functionality.
- Updated the command handling logic in the mod.rs file to reflect this change.
